move cairo interaction element to seperate file
