+++ stwo_cairo_prover/crates/cairo-air/src/builtins_air.rs
@@ -9,7 +9,7 @@ use stwo::prover::ComponentProver;
 use stwo_cairo_serialize::{CairoDeserialize, CairoSerialize};
 use stwo_constraint_framework::TraceLocationAllocator;
 
-use super::air::CairoInteractionElements;
+use crate::cairo_interaction_elements::CairoInteractionElements;
 use crate::components::{
     add_mod_builtin, bitwise_builtin, indented_component_display, mul_mod_builtin,
     pedersen_builtin, poseidon_builtin, range_check_builtin_bits_128, range_check_builtin_bits_96,
