+++ stwo_cairo_prover/crates/cairo-air/src/pedersen/air.rs
@@ -4,7 +4,7 @@ use stwo::prover::backend::simd::SimdBackend;
 use stwo::prover::ComponentProver;
 use stwo_constraint_framework::TraceLocationAllocator;
 
-use crate::air::CairoInteractionElements;
+use crate::cairo_interaction_elements::CairoInteractionElements;
 use crate::components::prelude::*;
 use crate::components::{indented_component_display, partial_ec_mul, pedersen_points_table};
 use crate::utils::{accumulate_relation_uses, RelationUsesDict};
