+++ stwo_cairo_prover/crates/cairo-air/src/poseidon/air.rs
@@ -4,7 +4,7 @@ use stwo::prover::backend::simd::SimdBackend;
 use stwo::prover::ComponentProver;
 use stwo_constraint_framework::TraceLocationAllocator;
 
-use crate::air::CairoInteractionElements;
+use crate::cairo_interaction_elements::CairoInteractionElements;
 use crate::components::prelude::*;
 use crate::components::{
     cube_252, indented_component_display, poseidon_3_partial_rounds_chain, poseidon_aggregator,
