+++ stwo_cairo_prover/crates/prover/src/witness/cairo.rs
@@ -1,9 +1,10 @@
 use std::array;
 
 use cairo_air::air::{
-    CairoClaim, CairoInteractionClaim, CairoInteractionElements, MemorySmallValue, PublicData,
-    PublicMemory, PublicSegmentRanges, SegmentRange,
+    CairoClaim, CairoInteractionClaim, MemorySmallValue, PublicData, PublicMemory,
+    PublicSegmentRanges, SegmentRange,
 };
+use cairo_air::cairo_interaction_elements::CairoInteractionElements;
 use itertools::Itertools;
 use stwo::core::fields::m31::M31;
 use stwo::prover::backend::simd::SimdBackend;
