+++ stwo_cairo_prover/crates/prover/src/witness/components/pedersen.rs
@@ -1,4 +1,4 @@
-use cairo_air::air::CairoInteractionElements;
+use cairo_air::cairo_interaction_elements::CairoInteractionElements;
 use cairo_air::pedersen::air::{
     Claim, InteractionClaim, PedersenContextClaim, PedersenContextInteractionClaim,
 };
