+++ stwo_cairo_prover/crates/prover/src/witness/opcodes.rs
@@ -1,4 +1,4 @@
-use cairo_air::air::CairoInteractionElements;
+use cairo_air::cairo_interaction_elements::CairoInteractionElements;
 use cairo_air::opcodes_air::{OpcodeClaim, OpcodeInteractionClaim};
 use stwo::prover::backend::simd::SimdBackend;
 use stwo_cairo_adapter::opcodes::StateTransitions;
